Repackaged core files
