+++ SEM/src/main/java/application/levelbuilder/LevelBuilder.java
@@ -1,16 +1,12 @@
 package application.levelbuilder;
 
-import application.Main;
-import application.core.*;
+import application.core.aliens.*;
 import org.lwjgl.input.Mouse;
 import org.newdawn.slick.*;
 import org.newdawn.slick.Color;
 import org.newdawn.slick.Graphics;
-import org.newdawn.slick.Image;
 import org.newdawn.slick.state.BasicGameState;
 import org.newdawn.slick.state.StateBasedGame;
-import org.newdawn.slick.state.transition.FadeInTransition;
-import org.newdawn.slick.state.transition.FadeOutTransition;
 
 import java.awt.*;
 import java.util.ArrayList;
