+++ src/main/java/application/controllers/LevelBuilder.java
@@ -5,12 +5,12 @@ import application.Main;
 import org.w3c.dom.Element;
 import java.util.ArrayList;
 import org.w3c.dom.Document;
-import application.core.Alien;
-import application.core.BigAlien;
-import application.core.MiniAlien;
-import application.core.FinalBoss;
-import application.core.SmallAlien;
-import application.core.MothershipAlien;
+import application.core.aliens.Alien;
+import application.core.aliens.BigAlien;
+import application.core.aliens.MiniAlien;
+import application.core.aliens.FinalBoss;
+import application.core.aliens.SmallAlien;
+import application.core.aliens.MothershipAlien;
 import org.newdawn.slick.Font;
 import org.newdawn.slick.Color;
 import org.newdawn.slick.Input;
