+++ src/main/java/application/core/Level.java
@@ -1,5 +1,7 @@
 package application.core;
 
+import application.core.aliens.Alien;
+
 import java.util.ArrayList;
 
 /**
