+++ src/main/java/application/core/projectiles/SmallProjectile.java
@@ -1,4 +1,4 @@
-package application.core;
+package application.core.projectiles;
 
 import application.Main;
