+++ src/main/java/application/core/upgrades/WeaponUpgrade.java
@@ -1,44 +1,44 @@
-package application.core;
+package application.core.upgrades;
 
 import application.Main;
 
 /**
- * Class for HealthUpgrade.
- * @author Ties Westerborg
+ * Class for WeaponUpgrade.
+ * @author Ties WesterBorg.
  */
 @SuppressWarnings({
         "checkstyle:magicnumber"
 })
-public final class HealthUpgrade extends Upgrade {
+public class WeaponUpgrade extends Upgrade {
 
     /**
-     * Constructor for HealthUpgrade.
+     * Constructor for WeaponUpgrade.
      * @param x x-coordinate.
      * @param y y-coordinate.
      */
-    public HealthUpgrade(int x, int y) {
+    public WeaponUpgrade(int x, int y) {
         tDirection = 1;
         tX = x;
         tY = y;
         tWidth = 50;
         tHeight = 50;
         tHealth = 1;
-        if (Main.DIFFICULTY == 1) {
+        if (tDifficulty == 1) {
             tSpeed = 2;
         }
-        else if (Main.DIFFICULTY == 2) {
-            tSpeed = 3;
+        else if (tDifficulty == 2) {
+            tSpeed = 2;
         }
         else {
-            tSpeed = 5;
+            tSpeed = 4;
         }
     }
 
     /**
-     * Getter method for the Image.
+     * Getter method for the belonging Image.
      * @return the belonging Image.
      */
     public org.newdawn.slick.Image getImage() {
-        return Main.UPGRADE_2;
+        return Main.UPGRADE_1;
     }
 }
