+++ src/test/java/application/core/BossProjectileTest.java
@@ -1,6 +1,7 @@
 package application.core;
 
 import application.Main;
+import application.core.projectiles.BossProjectile;
 import org.junit.Before;
 import org.junit.Test;
