+++ src/test/java/application/core/MiniAlienTest.java
@@ -1,6 +1,7 @@
 package application.core;
 
 import application.Main;
+import application.core.aliens.MiniAlien;
 import org.junit.Before;
 import org.junit.Test;
