+++ src/test/java/application/core/PlayerProjectileTest.java
@@ -1,6 +1,7 @@
 package application.core;
 
 import application.Main;
+import application.core.projectiles.PlayerProjectile;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
