+++ src/test/java/application/core/SmallProjectileTest.java
@@ -1,6 +1,7 @@
 package application.core;
 
 import application.Main;
+import application.core.projectiles.SmallProjectile;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
