+++ src/test/java/application/core/UpgradeTest.java
@@ -1,5 +1,7 @@
 package application.core;
 
+import application.core.upgrades.HealthUpgrade;
+import application.core.upgrades.Upgrade;
 import org.junit.Before;
 import org.junit.Test;
