+++ src/test/java/application/core/UpgradedProjectileTest.java
@@ -1,6 +1,7 @@
 package application.core;
 
 import application.Main;
+import application.core.projectiles.UpgradedProjectile;
 import org.junit.Before;
 import org.junit.Test;
